[REFACTOR] Refactor main.c
